--- eval/stillbox_eval/test_files_90.txt
@@ -1,192 +1,0 @@
-15/RGB_112_008.jpg
-15/RGB_178_002.jpg
-15/RGB_167_006.jpg
-15/RGB_153_007.jpg
-15/RGB_119_002.jpg
-15/RGB_135_003.jpg
-15/RGB_44_006.jpg
-15/RGB_32_002.jpg
-15/RGB_171_001.jpg
-15/RGB_114_009.jpg
-15/RGB_89_003.jpg
-15/RGB_197_009.jpg
-15/RGB_105_000.jpg
-15/RGB_72_004.jpg
-15/RGB_66_003.jpg
-15/RGB_25_007.jpg
-15/RGB_58_004.jpg
-15/RGB_28_003.jpg
-15/RGB_25_004.jpg
-15/RGB_140_003.jpg
-15/RGB_59_008.jpg
-15/RGB_19_001.jpg
-15/RGB_186_003.jpg
-15/RGB_113_009.jpg
-15/RGB_54_002.jpg
-15/RGB_130_003.jpg
-15/RGB_153_003.jpg
-15/RGB_103_007.jpg
-15/RGB_04_007.jpg
-15/RGB_110_008.jpg
-15/RGB_78_005.jpg
-15/RGB_26_005.jpg
-15/RGB_43_007.jpg
-15/RGB_190_003.jpg
-15/RGB_122_002.jpg
-15/RGB_102_008.jpg
-15/RGB_187_004.jpg
-15/RGB_03_005.jpg
-15/RGB_58_007.jpg
-15/RGB_37_004.jpg
-15/RGB_125_003.jpg
-15/RGB_190_002.jpg
-15/RGB_52_006.jpg
-15/RGB_37_005.jpg
-15/RGB_196_001.jpg
-15/RGB_53_003.jpg
-15/RGB_129_008.jpg
-15/RGB_74_003.jpg
-15/RGB_167_000.jpg
-15/RGB_195_002.jpg
-15/RGB_10_007.jpg
-15/RGB_131_003.jpg
-15/RGB_37_003.jpg
-15/RGB_38_009.jpg
-15/RGB_115_004.jpg
-15/RGB_91_008.jpg
-15/RGB_43_004.jpg
-15/RGB_187_005.jpg
-15/RGB_112_003.jpg
-15/RGB_19_002.jpg
-15/RGB_170_008.jpg
-15/RGB_17_000.jpg
-15/RGB_62_005.jpg
-15/RGB_148_004.jpg
-15/RGB_12_008.jpg
-15/RGB_169_004.jpg
-15/RGB_112_004.jpg
-15/RGB_71_001.jpg
-15/RGB_103_001.jpg
-15/RGB_178_005.jpg
-15/RGB_92_006.jpg
-15/RGB_40_009.jpg
-15/RGB_138_006.jpg
-15/RGB_146_005.jpg
-15/RGB_04_006.jpg
-15/RGB_02_008.jpg
-15/RGB_101_009.jpg
-15/RGB_103_009.jpg
-15/RGB_21_002.jpg
-15/RGB_144_008.jpg
-15/RGB_163_007.jpg
-15/RGB_06_001.jpg
-15/RGB_105_004.jpg
-15/RGB_199_009.jpg
-15/RGB_149_005.jpg
-15/RGB_63_008.jpg
-15/RGB_21_004.jpg
-15/RGB_03_002.jpg
-15/RGB_51_008.jpg
-15/RGB_110_001.jpg
-15/RGB_172_009.jpg
-15/RGB_158_005.jpg
-15/RGB_49_004.jpg
-15/RGB_173_008.jpg
-15/RGB_99_004.jpg
-15/RGB_24_001.jpg
-15/RGB_03_009.jpg
-15/RGB_41_009.jpg
-15/RGB_91_002.jpg
-15/RGB_132_001.jpg
-15/RGB_95_003.jpg
-15/RGB_167_005.jpg
-15/RGB_176_000.jpg
-15/RGB_142_008.jpg
-15/RGB_107_009.jpg
-15/RGB_122_005.jpg
-15/RGB_48_001.jpg
-15/RGB_103_005.jpg
-15/RGB_98_009.jpg
-15/RGB_162_001.jpg
-15/RGB_08_006.jpg
-15/RGB_169_002.jpg
-15/RGB_57_002.jpg
-15/RGB_86_004.jpg
-15/RGB_138_001.jpg
-15/RGB_05_005.jpg
-15/RGB_95_002.jpg
-15/RGB_28_002.jpg
-15/RGB_110_002.jpg
-15/RGB_102_002.jpg
-15/RGB_136_009.jpg
-15/RGB_28_007.jpg
-15/RGB_43_005.jpg
-15/RGB_39_006.jpg
-15/RGB_126_003.jpg
-15/RGB_62_001.jpg
-15/RGB_82_003.jpg
-15/RGB_75_008.jpg
-15/RGB_16_005.jpg
-15/RGB_94_005.jpg
-15/RGB_198_002.jpg
-15/RGB_90_001.jpg
-15/RGB_22_001.jpg
-15/RGB_90_000.jpg
-15/RGB_155_006.jpg
-15/RGB_124_007.jpg
-15/RGB_168_004.jpg
-15/RGB_96_008.jpg
-15/RGB_100_002.jpg
-15/RGB_131_008.jpg
-15/RGB_74_002.jpg
-15/RGB_141_007.jpg
-15/RGB_139_001.jpg
-15/RGB_102_005.jpg
-15/RGB_182_009.jpg
-15/RGB_37_002.jpg
-15/RGB_67_003.jpg
-15/RGB_60_001.jpg
-15/RGB_186_001.jpg
-15/RGB_171_002.jpg
-15/RGB_155_004.jpg
-15/RGB_50_008.jpg
-15/RGB_34_002.jpg
-15/RGB_132_003.jpg
-15/RGB_147_005.jpg
-15/RGB_99_008.jpg
-15/RGB_110_000.jpg
-15/RGB_114_008.jpg
-15/RGB_159_002.jpg
-15/RGB_76_007.jpg
-15/RGB_116_005.jpg
-15/RGB_67_002.jpg
-15/RGB_80_003.jpg
-15/RGB_30_000.jpg
-15/RGB_137_009.jpg
-15/RGB_130_002.jpg
-15/RGB_90_002.jpg
-15/RGB_34_008.jpg
-15/RGB_137_007.jpg
-15/RGB_45_001.jpg
-15/RGB_131_004.jpg
-15/RGB_06_000.jpg
-15/RGB_68_005.jpg
-15/RGB_104_008.jpg
-15/RGB_193_008.jpg
-15/RGB_182_000.jpg
-15/RGB_129_006.jpg
-15/RGB_107_005.jpg
-15/RGB_158_007.jpg
-15/RGB_192_001.jpg
-15/RGB_18_005.jpg
-15/RGB_90_009.jpg
-15/RGB_18_007.jpg
-15/RGB_94_000.jpg
-15/RGB_09_002.jpg
-15/RGB_94_001.jpg
-15/RGB_46_004.jpg
-15/RGB_126_000.jpg
-15/RGB_146_002.jpg
-15/RGB_161_006.jpg
-15/RGB_154_008.jpg
-15/RGB_94_003.jpg
