+++ models/SemPoseExpNet.py
@@ -1,0 +1,120 @@
+from re import S
+import torch
+import torch.nn as nn
+from torch import sigmoid
+from torch.nn.init import xavier_uniform_, zeros_
+
+
+def conv(in_planes, out_planes, kernel_size=3):
+    return nn.Sequential(
+        nn.Conv2d(in_planes, out_planes, kernel_size=kernel_size, padding=(kernel_size-1)//2, stride=2),
+        nn.ReLU(inplace=True)
+    )
+
+
+def upconv(in_planes, out_planes):
+    return nn.Sequential(
+        nn.ConvTranspose2d(in_planes, out_planes, kernel_size=4, stride=2, padding=1),
+        nn.ReLU(inplace=True)
+    )
+
+
+class SemPoseExpNet(nn.Module):
+
+    def __init__(self, nb_ref_imgs=2, output_exp=False):
+        super(SemPoseExpNet, self).__init__()
+        self.nb_ref_imgs = nb_ref_imgs
+        self.output_exp = output_exp
+
+        conv_planes = [16, 32, 64, 128, 256, 256, 256]
+        self.conv1 = conv(3*(1+self.nb_ref_imgs), conv_planes[0], kernel_size=7)
+        self.conv2 = conv(conv_planes[0], conv_planes[1], kernel_size=5)
+        self.conv3 = conv(conv_planes[1], conv_planes[2])
+        self.conv4 = conv(conv_planes[2], conv_planes[3])
+        self.conv5 = conv(conv_planes[3], conv_planes[4])
+        self.conv6 = conv(conv_planes[4], conv_planes[5])
+        self.conv7 = conv(conv_planes[5], conv_planes[6])
+        
+        conv_planes = [16, 32, 64, 128, 256, 256, 256]
+        self.sconv1 = conv(3*(1+self.nb_ref_imgs), conv_planes[0], kernel_size=7)
+        self.sconv2 = conv(conv_planes[0], conv_planes[1], kernel_size=5)
+        self.sconv3 = conv(conv_planes[1], conv_planes[2])
+        self.sconv4 = conv(conv_planes[2], conv_planes[3])
+        self.sconv5 = conv(conv_planes[3], conv_planes[4])
+        self.sconv6 = conv(conv_planes[4], conv_planes[5])
+        self.sconv7 = conv(conv_planes[5], conv_planes[6])
+
+        self.pose_pred = nn.Conv2d(2 * conv_planes[6], 6*self.nb_ref_imgs, kernel_size=1, padding=0)
+
+        if self.output_exp:
+            upconv_planes = [256, 128, 64, 32, 16]
+            self.upconv5 = upconv(conv_planes[4],   upconv_planes[0])
+            self.upconv4 = upconv(upconv_planes[0], upconv_planes[1])
+            self.upconv3 = upconv(upconv_planes[1], upconv_planes[2])
+            self.upconv2 = upconv(upconv_planes[2], upconv_planes[3])
+            self.upconv1 = upconv(upconv_planes[3], upconv_planes[4])
+
+            self.predict_mask4 = nn.Conv2d(upconv_planes[1], self.nb_ref_imgs, kernel_size=3, padding=1)
+            self.predict_mask3 = nn.Conv2d(upconv_planes[2], self.nb_ref_imgs, kernel_size=3, padding=1)
+            self.predict_mask2 = nn.Conv2d(upconv_planes[3], self.nb_ref_imgs, kernel_size=3, padding=1)
+            self.predict_mask1 = nn.Conv2d(upconv_planes[4], self.nb_ref_imgs, kernel_size=3, padding=1)
+
+    def init_weights(self):
+        for m in self.modules():
+            if isinstance(m, nn.Conv2d) or isinstance(m, nn.ConvTranspose2d):
+                xavier_uniform_(m.weight.data)
+                if m.bias is not None:
+                    zeros_(m.bias)
+
+    def forward(self, tgt_img, ref_imgs, sem_target_img, sem_ref_imgs):
+        assert(len(ref_imgs) == self.nb_ref_imgs) 
+        assert(len(sem_ref_imgs) == self.nb_ref_imgs)
+        
+        input = [tgt_img]
+        input.extend(ref_imgs)
+        input = torch.cat(input, 1)
+        out_conv1 = self.conv1(input)
+        out_conv2 = self.conv2(out_conv1)
+        out_conv3 = self.conv3(out_conv2)
+        out_conv4 = self.conv4(out_conv3)
+        out_conv5 = self.conv5(out_conv4)
+        out_conv6 = self.conv6(out_conv5)
+        out_conv7 = self.conv7(out_conv6)
+        
+        sem_input = [sem_ref_imgs]
+        sem_input.extend(sem_ref_imgs)
+        sem_input = torch.cat(sem_input, 1)
+        out_sconv1 = self.conv1(sem_input)
+        out_sconv2 = self.conv2(out_sconv1)
+        out_sconv3 = self.conv3(out_sconv2)
+        out_sconv4 = self.conv4(out_sconv3)
+        out_sconv5 = self.conv5(out_sconv4)
+        out_sconv6 = self.conv6(out_sconv5)
+        out_sconv7 = self.conv7(out_sconv6)
+
+        pose_input = torch.cat([out_conv7, out_sconv7])
+        pose = self.pose_pred(pose_input)
+        pose = pose.mean(3).mean(2)
+        pose = 0.01 * pose.view(pose.size(0), self.nb_ref_imgs, 6)
+
+        if self.output_exp:
+            out_upconv5 = self.upconv5(out_conv5  )[:, :, 0:out_conv4.size(2), 0:out_conv4.size(3)]
+            out_upconv4 = self.upconv4(out_upconv5)[:, :, 0:out_conv3.size(2), 0:out_conv3.size(3)]
+            out_upconv3 = self.upconv3(out_upconv4)[:, :, 0:out_conv2.size(2), 0:out_conv2.size(3)]
+            out_upconv2 = self.upconv2(out_upconv3)[:, :, 0:out_conv1.size(2), 0:out_conv1.size(3)]
+            out_upconv1 = self.upconv1(out_upconv2)[:, :, 0:input.size(2), 0:input.size(3)]
+
+            exp_mask4 = sigmoid(self.predict_mask4(out_upconv4))
+            exp_mask3 = sigmoid(self.predict_mask3(out_upconv3))
+            exp_mask2 = sigmoid(self.predict_mask2(out_upconv2))
+            exp_mask1 = sigmoid(self.predict_mask1(out_upconv1))
+        else:
+            exp_mask4 = None
+            exp_mask3 = None
+            exp_mask2 = None
+            exp_mask1 = None
+
+        if self.training:
+            return [exp_mask1, exp_mask2, exp_mask3, exp_mask4], pose
+        else:
+            return exp_mask1, pose
