+++ models/__init__.py
@@ -1,3 +1,4 @@
 from .DispNetS import DispNetS
 from .PoseExpNet import PoseExpNet
 from .SemDispNet import SemDispNetS
+from .SemPoseExpNet import SemPoseExpNet
