+++ utils/common.py
@@ -84,6 +84,31 @@ def log_output_tensorboard(writer, prefix, index, suffix, n_iter, depth, disp, w
             mask_to_show = tensor2array(mask[0, j], max_value=1, colormap='bone')
             writer.add_image('{} Exp mask Outputs {}'.format(prefix, whole_suffix), mask_to_show, n_iter)
 
+def log_output_tensorboard_semantic(
+    writer, prefix, index, suffix, n_iter, depth, disp, warped, diff, warped_sem, diff_sem, mask):
+    disp_to_show = tensor2array(disp[0], max_value=None, colormap='magma')
+    depth_to_show = tensor2array(depth[0], max_value=None)
+    writer.add_image('{} Dispnet Output Normalized{}/{}'.format(prefix, suffix, index), disp_to_show, n_iter)
+    writer.add_image('{} Depth Output Normalized{}/{}'.format(prefix, suffix, index), depth_to_show, n_iter)
+    # log warped images along with explainability mask
+    if (warped is None) or (diff is None):
+        return
+    for j, (warped_j, diff_j) in enumerate(zip(warped, diff)):
+        whole_suffix = '{} {}/{}'.format(suffix, j, index)
+        warped_to_show = tensor2array(warped_j)
+        diff_to_show = tensor2array(0.5*diff_j)
+        writer.add_image('{} Warped Outputs {}'.format(prefix, whole_suffix), warped_to_show, n_iter)
+        writer.add_image('{} Diff Outputs {}'.format(prefix, whole_suffix), diff_to_show, n_iter)
+        if mask is not None:
+            mask_to_show = tensor2array(mask[0, j], max_value=1, colormap='bone')
+            writer.add_image('{} Exp mask Outputs {}'.format(prefix, whole_suffix), mask_to_show, n_iter)
+    
+    for j, (warped_j, diff_j) in enumerate(zip(warped_sem, diff_sem)):
+        whole_suffix = '{} {}/{}'.format(suffix, j, index)
+        warped_to_show = tensor2array(warped_j)
+        diff_to_show = tensor2array(0.5*diff_j)
+        writer.add_image('{} Semantic Warped Outputs {}'.format(prefix, whole_suffix), warped_to_show, n_iter)
+        writer.add_image('{} Semantic Diff Outputs {}'.format(prefix, whole_suffix), diff_to_show, n_iter)
 
 def tensor2array(tensor, max_value=None, colormap='rainbow'):
     tensor = tensor.detach().cpu()
